+++ p_maputl.h
@@ -53,6 +53,9 @@ typedef struct {
 
 typedef boolean (*traverser_t)(intercept_t *in);
 
+// mbf21: search surrounding blockmap cells for a shootable target
+mobj_t *P_RoughTargetSearch(mobj_t *mo, angle_t fov, int distance);
+
 fixed_t P_AproxDistance (fixed_t dx, fixed_t dy);
 int     P_PointOnLineSide (fixed_t x, fixed_t y, line_t *line);
 int     P_PointOnDivlineSide (fixed_t x, fixed_t y, divline_t *line);